+++ library/src/main/java/io/github/fanky10/sociallogin/module/fragments/BaseFacebookLoginFragment.java
@@ -8,15 +8,8 @@ import com.facebook.GraphRequest;
 import com.facebook.GraphResponse;
 import com.facebook.login.LoginManager;
 import com.facebook.login.LoginResult;
-import com.twitter.sdk.android.core.Callback;
-import com.twitter.sdk.android.core.Result;
-import com.twitter.sdk.android.core.TwitterApiClient;
 import com.twitter.sdk.android.core.TwitterAuthConfig;
 import com.twitter.sdk.android.core.TwitterCore;
-import com.twitter.sdk.android.core.TwitterException;
-import com.twitter.sdk.android.core.TwitterSession;
-import com.twitter.sdk.android.core.identity.TwitterLoginButton;
-import com.twitter.sdk.android.core.models.User;
 
 import org.json.JSONObject;
 
@@ -24,9 +17,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.text.TextUtils;
-import android.util.Log;
-import android.view.View;
-import android.widget.Toast;
 
 import java.util.Arrays;
 
@@ -38,16 +28,11 @@ import io.github.fanky10.sociallogin.module.interfaces.ISocialLogin;
 /**
  * Created by carlospienovi1 on 12/4/15.
  */
-public abstract class BaseSocialLoginFragment extends Fragment
+public abstract class BaseFacebookLoginFragment extends Fragment
         implements ISocialLogin {
 
     private CallbackManager mFacebookCallbackManager;
     private IFacebook mIFacebook;
-    private TwitterLoginButton mTwitterLoginButton;
-
-    protected int getTwitterLoginButtonId() {
-        return -1;
-    }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -58,48 +43,6 @@ public abstract class BaseSocialLoginFragment extends Fragment
 
         TwitterAuthConfig authConfig = new TwitterAuthConfig(getString(R.string.twitter_app_key), getString(R.string.twitter_app_secret));
         Fabric.with(getContext(), new TwitterCore(authConfig));
-    }
-
-    @Override
-    public void onViewCreated(View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
-        mTwitterLoginButton = (TwitterLoginButton) view.findViewById(getTwitterLoginButtonId());
-
-        if (mTwitterLoginButton != null) {
-            mTwitterLoginButton.setCallback(new Callback<TwitterSession>() {
-                @Override
-                public void success(Result<TwitterSession> result) {
-                    TwitterSession session = result.data;
-
-                    final String userName = session.getUserName();
-                    final long userId = session.getUserId();
-                    String msg = "@" + userName + " logged in! (#" + userId + ")";
-                    Toast.makeText(getContext(), msg, Toast.LENGTH_LONG).show();
-
-                    TwitterApiClient twitterApiClient = TwitterCore.getInstance().getApiClient();
-                    twitterApiClient.getAccountService().verifyCredentials(false, false, new Callback<User>() {
-                        @Override
-                        public void success(Result<User> userResult) {
-                            String tFullName = userResult.data.name;
-                            String tProfileImage = userResult.data.profileImageUrl;
-                            Log.d("GATA", "User FullName: " + tFullName);
-                            Log.d("GATA", "User Profile: " + tProfileImage);
-                        }
-
-                        @Override
-                        public void failure(TwitterException e) {
-                            Toast.makeText(getContext(), "twitter login failed", Toast.LENGTH_SHORT).show();
-                        }
-                    });
-                }
-
-                @Override
-                public void failure(TwitterException e) {
-                    onSocialProviderConnectionFailure(e);
-                }
-            });
-        }
     }
 
     private FacebookCallback<LoginResult> getFacebookLoginCallback() {
@@ -149,12 +92,7 @@ public abstract class BaseSocialLoginFragment extends Fragment
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-
         mFacebookCallbackManager.onActivityResult(requestCode, resultCode, data);
-
-        if (mTwitterLoginButton != null) {
-            mTwitterLoginButton.onActivityResult(requestCode, resultCode, data);
-        }
     }
 
 }
